Horrible attempt at cleaning things up that probably just made it worse
