+++ src/libstd/sys/redox/ext/mod.rs
@@ -36,6 +36,7 @@ pub mod io;
 pub mod net;
 pub mod process;
 pub mod thread;
+pub(crate) mod unixsocket;
 
 /// A prelude for conveniently writing platform-specific code.
 ///
